external plugins: fix refresh race with plugin manager
PM refreshPlugins runs on EDT, causing a race with the EPM refresh on executor thread.

Since the event is already on the executor thread anyway, have EPM block
the event and run prior to the plugin manager.

This is to avoid plugins switching on/off rapidly when switching to a
new profile. Since most plugins default to on, PM will enable them and
EPM will want to uninstall them.

The inverse case of EPM installing a plugin which PM wants to stop is
not an issue because EPM cannot start plugins that are not enabled due
to the check in startPlugin()

(cherry picked from commit 89bfa828c90a0f5b1fbf8d5948bdd8d3c7139adc)