screenshot: imagecapture: add utility method for taking screenshots
The steps to take a screenshot are a little complex, so this adds a utility
method for plugins to take a screenshot. The screenshot plugin doesn't use this
because it has its overlay for the date/time which it sometimes uses instead.

(cherry picked from commit 09aa81dd139381aab88a6f86a486daa325532976)